examples/e2b-data-analyst/requirements.txt
blindfold-sdk>=1.3.0
e2b-code-interpreter>=1.0.0
openai>=1.0.0
python-dotenv>=1.0.0
